Completely new version
Over the past 2 days, I totally rewrote this site.  I used practices I've learned over the past 6 years as a professional developer.  I wanted this to be an easier end-user experience, as well as much easier to maintain as a dev.  Since this is a relatively small site, I decided to avoid using a database and just hardcode some things.  Yes, I know, but that's what I decided.  Maybe if I decide to add more and newer items to the site, I'll add a database, but that is more work than I wanted to put into this site right now.